GEODE-6904: Fix hang on locator when restarting due to lock (#3777)
GEM-2579 for support tracking.

Only hold the locatorLock while reading the locator variable in
restartWithDS. Holding this lock while recovering cluster configuration
was causing other messages to the locator, including the StartupMessage,
to block waiting for this lock.

In addition, wait forever for StartupMessage responses.
StartupOperation had a 15 second timeout, after which it would stop
waiting. That resulted in not having correct metadata about other
members in the system. We were also not waking up the thread in
waitForReplies when a StartupResponseMessage came in with a
rejectionMessage.

Co-Authored-By: Ernie Burghardt <eburghardt@pivotal.io>
Co-Authored-By: Dan Smith <dsmith@pivotal.io>

(cherry picked from commit 00ed2f3ca9025beb200a0851c530d041575bbd20)