GEM-2613: Reverting changes to wait forever for startup messages
The previous changes for GEM-2579 included a change to wait forever for
a response to a startup message, rather than moving on after 15 seconds.

It appears that the old behavior was covering up some bugs that lead to
hangs on startup if we don't give up after 15 seconds. Since the
customer fix for GEODE-2579 does not rely on waiting forever for startup
messages, reverting that part of the fix on this support branch only.