GEODE-7085: Ensure bitset is flushed in all code paths
In recordVersion, there was a code path where we would not call
flushBitSetDuringRecording before calling setVersionInBitSet. Moving the
flush to the top of the method to ensure that the bit set is always
flushed, and we never end up with a case where we try to set too large
of a bit.

By code inspection, we determined that initializeFrom was what got us
into the state that triggered this code path - it could leave
bitSetVersion at a small number while setting version to a large number.
Fixing initializeFrom so that it correctly sets bitSetVersion.

Co-Authored-By: Ernest Burghardt <eburghardt@pivotal.io>
(cherry picked from commit f17931bf541fc0255112f713931388d9ee0bbc30)