GEODE-7241 Refactor publish.gradle into -java and -war types (#4091)
Allows publication of Jar or War artifacts via an `apply` statement.
We then apply publish-war.gradle to our projects that build war
artifacts (and that we want to publish) and publish-java.gradle to the
jar libraries that we want to publish.

* Refactor publish.gradle into -java and -war types
* Use facets for test structures in geode-web-management

Replaces custom configuration and source-sets with the facets plug in, like we do elsewhere int he project
* geode-pulse is a WAR not a JAR. Fix its publication and expected pom.
* enable jar in build for test linking

(cherry picked from commit 7597643ab01c69038b18d4796077cfacc1dcb8f2)